GeminiScienceBot-B1/asd/requirements.txt
python-dotenv==1.0.1
google-generativeai==0.3.2
streamlit==1.30.0
pyttsx3
